working state new namespaces
