+++ TTT-WPFApp/MainWindow.xaml.cs
@@ -21,9 +21,107 @@ namespace TTT_WPFApp
     /// </summary>
     public partial class MainWindow : Window
     {
+        public TicTacToeGame myGame;
+        public Button[] buttons;
+
         public MainWindow()
         {
             InitializeComponent();
+
+            myGame = new TicTacToeGame();
+            buttons = new Button[9];
+
+            CreateButtons();
+            UpdateButtons();
+        }
+
+        private void CreateButtons()
+        {
+            int b = 0;
+            for (int i = 0; i < 3; i++)
+            {
+                for (int j = 0; j < 3; j++)
+                {
+                    Button button = new Button()
+                    {
+                        Content = "",
+                        Width = 80,
+                        Height = 80,
+                        FontSize = 36,
+                        Foreground = Brushes.White,
+                    };
+
+                    GameBoard.Children.Add(button);
+                    Grid.SetRow(button, i);
+                    Grid.SetColumn(button, j);
+
+                    button.Click += Button_Click;
+                    buttons[b++] = button;
+                }
+            }
+        }
+
+        private void UpdateButtons()
+        {
+            string[] flat = FlattenGrid(myGame.grid);
+            for (int i = 0; i < 9; i++)
+            {
+                buttons[i].Content = flat[i];
+                buttons[i].Background = GetBrushColour(flat[i]);
+            }
+        }
+
+        private Brush GetBrushColour(string text)
+        {
+            switch (text)
+            {
+                case "X":
+                    return Brushes.Red;
+                case "O":
+                    return Brushes.Blue;
+                default:
+                    return Brushes.White;
+            }
+        }
+
+        private string[] FlattenGrid(char[,] grid)
+        {
+            string[] flat = new string[9];
+
+            int p = 0;
+            for (int i = 0; i < 3; i++)
+            {
+                for (int j = 0; j < 3; j++)
+                {
+                    flat[p++] = grid[i, j].ToString();
+                }
+            }
+            return flat;
+        }
+
+        private string GetButtonText(int i)
+        {
+            return (string)buttons[i].Content;
+        }
+
+        private void ButtonClicked(int i)
+        {
+            if (GetButtonText(i) == "." && !myGame.GameOver())
+            {
+                myGame.PlaceMove(i / 3, i % 3);
+                UpdateButtons();
+            }
+        }
+
+        private void Button_Click(object sender, RoutedEventArgs e)
+        {
+            for (int i = 0; i < 9; i++)
+            {
+                if (sender == buttons[i])
+                {
+                    ButtonClicked(i);
+                }
+            }
         }
     }
 }
