+++ TTT-WPFApp/TTT-WPFApp.csproj
@@ -8,4 +8,8 @@
     <UseWPF>true</UseWPF>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\TTT-Library\TTT-Library.csproj" />
+  </ItemGroup>
+
 </Project>
