--- TicTacToeWPFUI/App.xaml
@@ -1,9 +1,0 @@
-﻿<Application x:Class="TicTacToeWPFUI.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:TicTacToeWPFUI"
-             StartupUri="MainWindow.xaml">
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
