--- TicTacToeWPFUI/MainWindow.xaml
@@ -1,27 +1,0 @@
-﻿<Window x:Class="TicTacToeWPFUI.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TicTacToeWPFUI"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800">
-
-    <Grid x:Name="GameBoard"  Width="300" Height="300">
-
-        <Grid.RowDefinitions>
-            <RowDefinition Height="100"/>
-            <RowDefinition Height="100"/>
-            <RowDefinition Height="100"/>
-        </Grid.RowDefinitions>
-        
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="100"/>
-            <ColumnDefinition Width="100"/>
-            <ColumnDefinition Width="100"/>
-        </Grid.ColumnDefinitions>
-
-        
-
-    </Grid>
-</Window>
