--- TicTacToeWinFormsUI/TicTacToeWinFormsUI.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>WinExe</OutputType>
-    <TargetFramework>net6.0-windows</TargetFramework>
-    <Nullable>enable</Nullable>
-    <UseWindowsForms>true</UseWindowsForms>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\TicTacToeLibrary\TicTacToeLibrary.csproj" />
-  </ItemGroup>
-
-</Project>
